requirements.txt
Flask>=2.2.0
sqlite3
pyngrok>=5.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
matplotlib>=3.5.0
Faker>=15.0.0
gunicorn>=20.1.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
plotly>=5.0.0
requests>=2.28.0
werkzeug>=2.2.0
